refactor: Post DTO, Entity, Mapper 확인 및 점검
1. MapStruct 적용
