+++ src/main/java/com/bluehair/hanghaefinalproject/post/mapper/PostMapStruct.java
@@ -1,6 +1,7 @@
 package com.bluehair.hanghaefinalproject.post.mapper;
 
-import com.bluehair.hanghaefinalproject.post.dto.serviceDto.MainPostDto;
+import com.bluehair.hanghaefinalproject.post.dto.responseDto.ResponseInfoPostDto;
+import com.bluehair.hanghaefinalproject.post.dto.responseDto.ResponseMainPostDto;
 import com.bluehair.hanghaefinalproject.post.dto.serviceDto.MainProfileDto;
 import com.bluehair.hanghaefinalproject.post.dto.serviceDto.PostDto;
 import com.bluehair.hanghaefinalproject.post.entity.Post;
@@ -14,6 +15,8 @@ import java.util.List;
 public interface PostMapStruct {
     PostMapStruct POST_MAPPER = Mappers.getMapper(PostMapStruct.class);
     Post PostDtoToPost(PostDto postDto, String nickname);
-    MainPostDto PostToMainPostDto(Long id, String title,String postImg ,Long likeCount, Long viewCount, String musicFile
-                                ,List<String> tagList,List<MainProfileDto> mainProfileList, boolean isLiked);
+    ResponseMainPostDto PostToMainPostDto(Long id, String title, String postImg , Long likeCount, Long viewCount, String musicFile
+                                , List<String> tagList, List<MainProfileDto> mainProfileList, boolean isLiked);
+
+    ResponseInfoPostDto postToResponseInfoPostDto(Post post, Boolean isLiked);
 }
